Split out computeProductsWithPrices
